Fix minor bugs
Apply hot fix in medium.py. Sampling scalar fields with inf values occasionally (with a low probability) returns a NaN value.
This issue is problematic when modelling occlusions (with infinite attenuation) in the scene and we need to sample the attenuation coefficient in object space.

Ensure the number of maximum ray steps is an integer in pyTorchRayTrace.py.
This number should not affect typical use cases because all the rays should have exited the scene earlier than this maximum number of steps and stopped the tracing.

Eliminate zero entries in the sparse projection matrix to save GPU memory in pyTorchAlgebraicPropagation.py.

Make figure save path in displaygrayscale.py more robust.